#refactor

+++ src/main/java/compute/model/FeeCalculationRequest.java
@@ -22,7 +22,7 @@ public class FeeCalculationRequest {
     // execution details
     private String orderExecutionId;
     private String ticketId;
-    private Double quantity;
+    private Integer quantity;
     private Double price;
     private String exchangeMIC;
 
@@ -126,11 +126,11 @@ public class FeeCalculationRequest {
         this.ticketId = ticketId;
     }
 
-    public Double getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
